+++ issuer-service/src/test/java/ch/admin/bj/swiyu/issuer/oid4vci/service/CredentialServiceTest.java
@@ -6,11 +6,11 @@
 
 package ch.admin.bj.swiyu.issuer.oid4vci.service;
 
+import ch.admin.bj.swiyu.issuer.api.oid4vci.CredentialEndpointRequestDto;
 import ch.admin.bj.swiyu.issuer.api.oid4vci.CredentialEnvelopeDto;
-import ch.admin.bj.swiyu.issuer.api.oid4vci.CredentialRequestDto;
-import ch.admin.bj.swiyu.issuer.api.oid4vci.DeferredCredentialRequestDto;
+import ch.admin.bj.swiyu.issuer.api.oid4vci.DeferredCredentialEndpointRequestDto;
 import ch.admin.bj.swiyu.issuer.api.oid4vci.OAuthTokenDto;
-import ch.admin.bj.swiyu.issuer.api.oid4vci.issuance_v2.CredentialRequestDtoV2;
+import ch.admin.bj.swiyu.issuer.api.oid4vci.issuance_v2.CredentialEndpointRequestDtoV2;
 import ch.admin.bj.swiyu.issuer.api.oid4vci.issuance_v2.ProofsDto;
 import ch.admin.bj.swiyu.issuer.common.config.ApplicationProperties;
 import ch.admin.bj.swiyu.issuer.common.exception.CredentialRequestError;
@@ -272,7 +272,7 @@ class CredentialServiceTest {
                 new CredentialOfferMetadata(true, null, null, null),
                 UUID.randomUUID());
 
-        DeferredCredentialRequestDto deferredRequest = new DeferredCredentialRequestDto(credentialOffer.getTransactionId());
+        DeferredCredentialEndpointRequestDto deferredRequest = new DeferredCredentialEndpointRequestDto(credentialOffer.getTransactionId());
 
         when(credentialOfferRepository.findByAccessToken(accessToken)).thenReturn(Optional.of(credentialOffer));
         when(credentialOfferRepository.findByPreAuthorizedCode(any(UUID.class))).thenReturn(Optional.empty());
@@ -301,7 +301,7 @@ class CredentialServiceTest {
                 new CredentialOfferMetadata(true, null, null, null),
                 UUID.randomUUID());
 
-        DeferredCredentialRequestDto deferredRequest = new DeferredCredentialRequestDto(credentialOffer.getTransactionId());
+        DeferredCredentialEndpointRequestDto deferredRequest = new DeferredCredentialEndpointRequestDto(credentialOffer.getTransactionId());
 
         when(credentialOfferRepository.findByAccessToken(accessToken)).thenReturn(Optional.of(credentialOffer));
         when(credentialOfferRepository.findByPreAuthorizedCode(any(UUID.class))).thenReturn(Optional.empty());
@@ -321,7 +321,7 @@ class CredentialServiceTest {
         UUID transactionId = UUID.randomUUID();
         UUID accessToken = UUID.randomUUID();
 
-        DeferredCredentialRequestDto deferredRequest = new DeferredCredentialRequestDto(transactionId);
+        DeferredCredentialEndpointRequestDto deferredRequest = new DeferredCredentialEndpointRequestDto(transactionId);
 
         CredentialOffer credentialOffer = getCredentialOffer(
                 CredentialStatusType.DEFERRED,
@@ -349,7 +349,7 @@ class CredentialServiceTest {
         UUID transactionId = UUID.randomUUID();
         UUID accessToken = UUID.randomUUID();
 
-        DeferredCredentialRequestDto deferredRequest = new DeferredCredentialRequestDto(transactionId);
+        DeferredCredentialEndpointRequestDto deferredRequest = new DeferredCredentialEndpointRequestDto(transactionId);
 
         CredentialOffer credentialOffer = spy(getCredentialOffer(
                 CredentialStatusType.READY,
@@ -394,7 +394,7 @@ class CredentialServiceTest {
         UUID transactionId = UUID.randomUUID();
         UUID accessToken = UUID.randomUUID();
 
-        DeferredCredentialRequestDto deferredRequest = new DeferredCredentialRequestDto(transactionId);
+        DeferredCredentialEndpointRequestDto deferredRequest = new DeferredCredentialEndpointRequestDto(transactionId);
 
         CredentialOffer credentialOffer = spy(getCredentialOffer(
                 CredentialStatusType.READY,
@@ -484,7 +484,7 @@ class CredentialServiceTest {
     @Test
     void createCredentialV2_deferred_thenSuccess() throws JsonProcessingException {
         // Arrange
-        CredentialRequestDtoV2 requestDto = mock(CredentialRequestDtoV2.class);
+        CredentialEndpointRequestDtoV2 requestDto = mock(CredentialEndpointRequestDtoV2.class);
         UUID accessToken = UUID.randomUUID();
 
         CredentialRequestClass credentialRequest = mock(CredentialRequestClass.class);
@@ -515,7 +515,7 @@ class CredentialServiceTest {
     @Test
     void createCredentialV2_nonDeferred_thenSuccess() {
         // Arrange
-        CredentialRequestDtoV2 requestDto = mock(CredentialRequestDtoV2.class);
+        CredentialEndpointRequestDtoV2 requestDto = mock(CredentialEndpointRequestDtoV2.class);
         UUID accessToken = UUID.randomUUID();
 
         CredentialRequestClass credentialRequest = mock(CredentialRequestClass.class);
@@ -558,7 +558,7 @@ class CredentialServiceTest {
         UUID accessToken = UUID.randomUUID();
         UUID transactionId = UUID.randomUUID();
         var expirationTimeStamp = now().plusSeconds(1000).getEpochSecond();
-        DeferredCredentialRequestDto deferredRequest = new DeferredCredentialRequestDto(transactionId);
+        DeferredCredentialEndpointRequestDto deferredRequest = new DeferredCredentialEndpointRequestDto(transactionId);
         var offer = getCredentialOffer(CredentialStatusType.IN_PROGRESS, expirationTimeStamp, offerData, accessToken, UUID.randomUUID(), UUID.randomUUID(), null, UUID.randomUUID());
 
         when(credentialOfferRepository.findByAccessToken(accessToken)).thenReturn(Optional.of(offer));
@@ -574,7 +574,7 @@ class CredentialServiceTest {
         UUID accessToken = UUID.randomUUID();
         UUID transactionId = UUID.randomUUID();
         var expirationTimeStamp = now().minusSeconds(1).getEpochSecond();
-        DeferredCredentialRequestDto deferredRequest = new DeferredCredentialRequestDto(transactionId);
+        DeferredCredentialEndpointRequestDto deferredRequest = new DeferredCredentialEndpointRequestDto(transactionId);
         var offer = getCredentialOffer(CredentialStatusType.READY, expirationTimeStamp, offerData, accessToken, UUID.randomUUID(), UUID.randomUUID(), null, transactionId);
 
         when(credentialOfferRepository.findByAccessToken(accessToken)).thenReturn(Optional.of(offer));
@@ -591,7 +591,7 @@ class CredentialServiceTest {
         UUID accessToken = UUID.randomUUID();
         UUID transactionId = UUID.randomUUID();
         var expirationTimeStamp = now().plusSeconds(100).getEpochSecond();
-        CredentialRequestDto credentialRequestDto = getCredentialRequestDto();
+        CredentialEndpointRequestDto credentialRequestDto = getCredentialRequestDto();
         var offer = getCredentialOffer(status, expirationTimeStamp, offerData, accessToken, UUID.randomUUID(), UUID.randomUUID(), null, transactionId);
 
         when(credentialOfferRepository.findByAccessToken(accessToken)).thenReturn(Optional.of(offer));
@@ -606,7 +606,7 @@ class CredentialServiceTest {
         UUID accessToken = UUID.randomUUID();
         UUID transactionId = UUID.randomUUID();
         var expirationTimeStamp = now().plusSeconds(100).getEpochSecond();
-        CredentialRequestDtoV2 credentialRequestDto = getCredentialRequestDtoV2("not-test", null);
+        CredentialEndpointRequestDtoV2 credentialRequestDto = getCredentialRequestDtoV2("not-test", null);
         var offer = getCredentialOffer(CredentialStatusType.IN_PROGRESS, expirationTimeStamp, offerData, accessToken, UUID.randomUUID(), UUID.randomUUID(), null, transactionId);
         var config = mock(CredentialConfiguration.class);
         when(config.getFormat()).thenReturn("vc+sd-jwt");
@@ -660,16 +660,16 @@ class CredentialServiceTest {
                 .build();
     }
 
-    private @NotNull CredentialRequestDto getCredentialRequestDto() {
-        return new CredentialRequestDto(
+    private @NotNull CredentialEndpointRequestDto getCredentialRequestDto() {
+        return new CredentialEndpointRequestDto(
                 "vc+sd-jwt",
                 new HashMap<>(),
                 null
         );
     }
 
-    private @NotNull CredentialRequestDtoV2 getCredentialRequestDtoV2(String credentialConfigurationId, ProofsDto proofs) {
-        return new CredentialRequestDtoV2(
+    private @NotNull CredentialEndpointRequestDtoV2 getCredentialRequestDtoV2(String credentialConfigurationId, ProofsDto proofs) {
+        return new CredentialEndpointRequestDtoV2(
                 credentialConfigurationId,
                 proofs,
                 null