+++ CHANGELOG.md
@@ -30,6 +30,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Changed the `didresolver` version from 2.0.1 to 2.1.3.
 - Updated ApiErrorDto and reused it for every error response. This allows for a more consistent error
   response structure.
+- Rename of
+    - `CreateCredentialRequestDto` to `CredentialEndpointRequestDto` (without dto in openapi schema name)
+    - `CredentialRequestDtoV2` to `CredentialEndpointRequestDtoV2` (without dto in openapi schema name)
+    - `CredentialResponseDto` to `CredentialEndpointResponseDto` (without dto in openapi schema name)
+    - `CredentialResponseDtoV2` to `CredentialEndpointResponseDtoV2` (without dto in openapi schema name)
+      to fix inconsistent openapi definition.
 
 ### Fixed